renamed `clients` dir to `service-clients`
