+++ src/service-clients/asset-client.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const Client = require('./base-client')
+const Client = require('./client')
 const { uploadFile } = require('../utils/shark-upload-file')
 const mime = require('mime/lite')
