+++ src/service-clients/consent-client.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const Client = require('./base-client')
+const Client = require('./client')
 
 class ConsentClient {
   constructor (url, options = {}) {
